refactor: 활성 이벤트 조회메서드 organizationService -> eventService로 변경 및 사용하지 않는 코드 삭제
