+++ server/src/test/java/com/ahmadda/application/OrganizationServiceTest.java
@@ -4,11 +4,8 @@ import com.ahmadda.annotation.IntegrationTest;
 import com.ahmadda.application.dto.LoginMember;
 import com.ahmadda.application.dto.OrganizationCreateRequest;
 import com.ahmadda.application.dto.OrganizationUpdateRequest;
-import com.ahmadda.common.exception.ForbiddenException;
 import com.ahmadda.common.exception.NotFoundException;
 import com.ahmadda.common.exception.UnprocessableEntityException;
-import com.ahmadda.domain.event.Event;
-import com.ahmadda.domain.event.EventOperationPeriod;
 import com.ahmadda.domain.event.EventRepository;
 import com.ahmadda.domain.member.Member;
 import com.ahmadda.domain.member.MemberRepository;
@@ -119,68 +116,6 @@ class OrganizationServiceTest {
     }
 
     @Test
-    void 존재하지_않는_조직의_이벤트를_조회하면_예외가_발생한다() {
-        // given
-        var member = memberRepository.save(Member.create("user", "user@test.com", "testPicture"));
-        var loginMember = new LoginMember(member.getId());
-
-        // when // then
-        assertThatThrownBy(() -> sut.getOrganizationEvents(999L, loginMember))
-                .isInstanceOf(NotFoundException.class)
-                .hasMessage("존재하지 않는 조직입니다.");
-    }
-
-    @Test
-    void 여러_조직의_이벤트가_있을때_선택된_조직의_활성화된_이벤트만_가져온다() {
-        // given
-        var member = memberRepository.save(Member.create("name", "test@test.com", "testPicture"));
-        var loginMember = new LoginMember(member.getId());
-        var orgA = organizationRepository.save(createOrganization("OrgA", "DescA", "a.png"));
-        var orgB = organizationRepository.save(createOrganization("OrgB", "DescB", "b.png"));
-        var orgMemberA =
-                organizationMemberRepository.save(OrganizationMember.create(
-                        "nickname",
-                        member,
-                        orgA,
-                        OrganizationMemberRole.USER
-                ));
-        var orgMemberB =
-                organizationMemberRepository.save(OrganizationMember.create(
-                        "nickname",
-                        member,
-                        orgB,
-                        OrganizationMemberRole.USER
-                ));
-
-        var now = LocalDateTime.now();
-        eventRepository.save(createEvent(orgMemberA, orgA, "EventA1", now.plusDays(1), now.plusDays(2)));
-        eventRepository.save(createEvent(orgMemberA, orgA, "EventA2", now.plusDays(2), now.plusDays(3)));
-        eventRepository.save(createEvent(orgMemberA, orgA, "EventA3", now.minusDays(2), now.minusDays(1))); // inactive
-        eventRepository.save(createEvent(orgMemberB, orgB, "EventB1", now.plusDays(1), now.plusDays(2)));
-
-        // when
-        var events = sut.getOrganizationEvents(orgA.getId(), loginMember);
-
-        // then
-        assertThat(events).hasSize(2)
-                .extracting(Event::getTitle)
-                .containsExactlyInAnyOrder("EventA1", "EventA2");
-    }
-
-    @Test
-    void 조직원이_아니면_조직의_이벤트를_조회시_예외가_발생한다() {
-        // given
-        var member = memberRepository.save(Member.create("user", "user@test.com", "testPicture"));
-        var organization = organizationRepository.save(createOrganization("Org", "Desc", "img.png"));
-        var loginMember = new LoginMember(member.getId());
-
-        // when // then
-        assertThatThrownBy(() -> sut.getOrganizationEvents(organization.getId(), loginMember))
-                .isInstanceOf(ForbiddenException.class)
-                .hasMessage("조직에 참여하지 않아 권한이 없습니다.");
-    }
-
-    @Test
     void 초대코드를_통해_조직에_참여할_수_있다() {
         // given
         var member1 = memberRepository.save(Member.create("user1", "user1@test.com", "testPicture"));
@@ -269,29 +204,6 @@ class OrganizationServiceTest {
         assertThatThrownBy(() -> sut.participateOrganization(organization.getId(), loginMember, request))
                 .isInstanceOf(UnprocessableEntityException.class)
                 .hasMessage("잘못된 초대코드입니다.");
-    }
-
-    @Test
-    void DEPRECATED_항상_우아한코스_조직을_반환한다() {
-        // when
-        var woowacourse = sut.alwaysGetWoowacourse();
-
-        // then
-        assertThat(woowacourse.getName()).isEqualTo(OrganizationService.WOOWACOURSE_NAME);
-    }
-
-    @Test
-    void DEPRECATED_여러번_요청해도_항상_우아한코스_조직을_반환한다() {
-        //given
-        Organization woowacourse =
-                Organization.create(OrganizationService.WOOWACOURSE_NAME, "우아한테크코스입니당딩동", "imageUrl");
-        organizationRepository.save(woowacourse);
-
-        // when
-        var getWoowacourse = sut.alwaysGetWoowacourse();
-
-        // then
-        assertThat(getWoowacourse.getName()).isEqualTo(OrganizationService.WOOWACOURSE_NAME);
     }
 
     @Test
@@ -416,29 +328,6 @@ class OrganizationServiceTest {
 
     private Organization createOrganization(String name, String description, String imageUrl) {
         return Organization.create(name, description, imageUrl);
-    }
-
-    private Event createEvent(
-            OrganizationMember organizer,
-            Organization organization,
-            String title,
-            LocalDateTime start,
-            LocalDateTime end
-    ) {
-
-        return Event.create(
-                title,
-                "description",
-                "place",
-                organizer,
-                organization,
-                EventOperationPeriod.create(
-                        start, end,
-                        end.plusHours(1), end.plusHours(2),
-                        start.minusDays(1)
-                ),
-                100
-        );
     }
 
     private OrganizationCreateRequest createOrganizationCreateRequest(
